# openapi.json
{
  "components": {
    "schemas": {
...
      "AuthSession": {
        "additionalProperties": false,
        "properties": {
          "client_id": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "expires_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "revoked_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "client_id",
          "created_at"
        ],
        "type": "object"
      },
...
      "Client": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "created_at"
        ],
        "type": "object"
      },
...
      "CreateClientRequest": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
      "Error": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "reason"
        ],
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "StatusResponse": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "consoleSession": {
        "in": "cookie",
        "name": "wingman_session",
        "type": "apiKey"
      },
      "rootBearer": {
        "description": "Daemon root credential",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Wingman API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/sessions": {
      "get": {
        "operationId": "listAuthSessions",
        "parameters": [
          {
            "description": "Client identity for resource attribution and scoping",
            "in": "header",
            "name": "X-Wingman-Client",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Client identity",
            "in": "query",
            "name": "client_id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuthSession"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request failed"
          }
        },
        "security": [
          {
            "rootBearer": []
          }
        ],
        "summary": "List auth sessions"
      }
    },
    "/auth/sessions/{id}": {
      "delete": {
        "operationId": "revokeAuthSession",
        "parameters": [
          {
            "description": "Client identity for resource attribution and scoping",
            "in": "header",
            "name": "X-Wingman-Client",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request failed"
          }
        },
        "security": [
          {
            "rootBearer": []
          }
        ],
        "summary": "Revoke an auth session"
      }
    },
...
    "/clients": {
      "get": {
        "operationId": "listClients",
        "parameters": [
          {
            "description": "Client identity for resource attribution and scoping",
            "in": "header",
            "name": "X-Wingman-Client",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Client"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request failed"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "consoleSession": []
          }
        ],
        "summary": "List API clients"
      },
      "post": {
        "operationId": "createClient",
        "parameters": [
          {
            "description": "Client identity for resource attribution and scoping",
            "in": "header",
            "name": "X-Wingman-Client",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClientRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request failed"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "consoleSession": []
          }
        ],
        "summary": "Create an API client"
      }
    },
    "/clients/{id}": {
      "get": {
        "operationId": "getClient",
        "parameters": [
          {
            "description": "Client identity for resource attribution and scoping",
            "in": "header",
            "name": "X-Wingman-Client",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request failed"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "consoleSession": []
          }
        ],
        "summary": "Get an API client"
      }
    },
...
  }
}